migration down fixes

+++ app/database/migrations/2015_08_11_150407_create_armadas_table.php
@@ -28,7 +28,10 @@ class CreateArmadasTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('armadas');
+		if (Schema::hasTable('armadas'))
+		{
+			Schema::drop('armadas');
+		}
 	}
 
 }
