+++ app/database/migrations/2016_05_20_173228_create_armada_membership_requests_table.php
@@ -36,7 +36,10 @@ class CreateArmadaMembershipRequestsTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('armada_membership_requests');
+		if (Schema::hasTable('armada_membership_requests'))
+		{
+			Schema::drop('armada_membership_requests');
+		}
 	}
 
 }
