+++ app/database/migrations/2016_05_26_103917_create_fb_apprequests_table.php
@@ -42,7 +42,10 @@ class CreateFbApprequestsTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('fb_apprequests');
+		if (Schema::hasTable('fb_apprequests'))
+		{
+			Schema::drop('fb_apprequests');
+		}
 	}
 
 }
