+++ app/database/migrations/2016_06_03_130750_create_users_energy_table.php
@@ -37,7 +37,10 @@ class CreateUsersEnergyTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('users_energy');
+		if (Schema::hasTable('users_energy'))
+		{
+			Schema::drop('users_energy');
+		}
 	}
 
 }
