+++ app/database/migrations/2016_06_11_142206_create_messages_table.php
@@ -34,7 +34,10 @@ class CreateMessagesTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('messages');
+		if (Schema::hasTable('messages'))
+		{
+			Schema::drop('messages');
+		}
 	}
 
 }
